Enum riff_wave::ReadErrorKind

pub enum ReadErrorKind {
    NotARiffFile,
    NotAWaveFile,
    NotAnUncompressedPcmWaveFile(u16),
    FmtChunkTooShort,
    NumChannelsIsZero,
    SampleRateIsZero,
    UnsupportedBitsPerSample(u16),
    InvalidBitsPerSample(u16, u16),
}

Represents a file format error, when the wave file is incorrect or unsupported.

Variants

NotARiffFile

The file does not start with a "RIFF" tag and chunk size.

NotAWaveFile

The file doesn't continue with "WAVE" after the RIFF chunk header.

NotAnUncompressedPcmWaveFile

This file is not an uncompressed PCM wave file. Only uncompressed files are supported.

FmtChunkTooShort

This file is missing header data and can't be parsed.

NumChannelsIsZero

The number of channels is zero, which is invalid.

SampleRateIsZero

The sample rate is zero, which is invalid.

UnsupportedBitsPerSample

Only 8-bit, 16-bit, 24-bit and 32-bit PCM files are supported.

InvalidBitsPerSample

We don't currently support extended PCM wave files where the actual bits per sample is less than the container size.